Scripts/Spells: Small cleanup for generic spell scripts
