+++ src/server/scripts/EasternKingdoms/SunwellPlateau/boss_muru.cpp
@@ -608,6 +608,31 @@ class spell_summon_blood_elves_periodic : public AuraScript
     }
 };
 
+// 46284 - Negative Energy Periodic
+class spell_muru_negative_energy_periodic : public AuraScript
+{
+    PrepareAuraScript(spell_muru_negative_energy_periodic);
+
+    bool Validate(SpellInfo const* spellInfo) override
+    {
+        return ValidateSpellInfo({ spellInfo->GetEffect(EFFECT_0).TriggerSpell });
+    }
+
+    void PeriodicTick(AuraEffect const* aurEff)
+    {
+        PreventDefaultAction();
+
+        CastSpellExtraArgs args(aurEff);
+        args.AddSpellMod(SPELLVALUE_MAX_TARGETS, aurEff->GetTickNumber() / 10 + 1);
+        GetTarget()->CastSpell(nullptr, aurEff->GetSpellEffectInfo().TriggerSpell, args);
+    }
+
+    void Register() override
+    {
+        OnEffectPeriodic += AuraEffectPeriodicFn(spell_muru_negative_energy_periodic::PeriodicTick, EFFECT_0, SPELL_AURA_PERIODIC_TRIGGER_SPELL);
+    }
+};
+
 void AddSC_boss_muru()
 {
     RegisterSunwellPlateauCreatureAI(boss_muru);
@@ -621,4 +646,5 @@ void AddSC_boss_muru()
     RegisterSpellScript(spell_dark_fiend_skin);
     RegisterSpellScript(spell_transform_visual_missile_periodic);
     RegisterSpellScript(spell_summon_blood_elves_periodic);
+    RegisterSpellScript(spell_muru_negative_energy_periodic);
 }
